Rearranged to isolate ComplexTimeSeries components that caused some collisions.  Better organization resulted anyway
